gca/pyproject.toml
[project]
name = "gca"
version = "0.1.0"
description = "GCA Model for UI Vision"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "scipy",
    "pillow",
    "datasets",
    "huggingface_hub",
    "tqdm",
    "matplotlib>=3.10.8",
    "safetensors",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu130", marker = "sys_platform != 'darwin'" },
]
torchvision = [
    { index = "pytorch-cu130", marker = "sys_platform != 'darwin'" },
]
